image-pipeline/image-pipeline/milvus-init/requirements.txt
pymilvus==2.3.3
requests==2.31.0
Pillow==10.1.0
